Changed requirement for tracker to nonsensical input by throwing exceptions instead of ignoring the method calls.
