+++ src/main/java/pathtracker/doubles/TrackerStub.java
@@ -33,7 +33,7 @@ public class TrackerStub implements Tracker {
     }
 
     @Override
-    public void addCharacter(String name, CharacterType characterType, int initiative) {
+    public void addCharacter(String name, CharacterType type, int initiative) {
 
     }
